Implementation of Policies and converting Tour controller to work with admin side

+++ routes/admin.php
@@ -22,8 +22,14 @@ Route::middleware(['jwt.auth', 'role:superadmin|admin'])->group(function () {
     Route::delete('users/{user}', 'Admin\MobileUserController@destroy')->name('admin.users.destroy');
 
     Route::get('admins', 'Admin\AdminController@index')->name('admin.admins.index');
-    Route::post('admins', 'Admin\AdminController@store')->name('admin.admins.store');
+    Route::post('admins', 'Admin\AdminController@store')->name('admin.admins.store')->middleware(['can:create,App\Admin']);
     Route::get('admins/{admin}', 'Admin\AdminController@show')->name('admin.admins.show');
-    Route::patch('admins/{admin}', 'Admin\AdminController@update')->name('admin.admins.update');
-    Route::delete('admins/{admin}', 'Admin\AdminController@destroy')->name('admin.admins.destroy');
+    Route::patch('admins/{admin}', 'Admin\AdminController@update')->name('admin.admins.update')->middleware(['can:update,admin']);
+    Route::delete('admins/{admin}', 'Admin\AdminController@destroy')->name('admin.admins.destroy')->middleware(['can:delete,admin']);
+
+    Route::resource('tours', 'Admin\TourController', ['as' => 'admin']);
+    Route::put('tours/{tour}/media', 'Admin\TourController@uploadMedia')->name('admin.tours.media');
+    // Route::resource('tours/{tour}/stops', 'Admin\StopController', ['as' => 'admin']);
+    // Route::put('tours/{tour}/stops/{stop}/order', 'Admin\StopController@changeOrder')->name('admin.stops.order');
+    // Route::put('tours/{tour}/stops/{stop}', 'Admin\StopController@uploadMedia')->name('admin.stops.media');
 });
